Consolidate LoadDatabase scripts into single Bean
