+++ src/main/java/com/ams/restapi/attendance/AttendanceRepository.java
@@ -2,7 +2,6 @@ package com.ams.restapi.attendance;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.List;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -10,8 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-interface AttendanceRepository extends JpaRepository<AttendanceLog, Long> {
-
+public interface AttendanceRepository extends JpaRepository<AttendanceLog, Long> {
     @Query("SELECT a FROM AttendanceLog a WHERE "
     + "(:room is null or a.room = :room) and "
     + "(:date is null or a.date = :date) and "
