+++ src/main/java/com/ams/restapi/timeConfig/TimeRepository.java
@@ -4,5 +4,6 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface TimeRepository extends JpaRepository<TimeConfig, Long> {
+public interface TimeRepository extends JpaRepository<TimeConfig, Long> {
+    
 }
